+++ Sources/VCore/Global Functions/KeyPath Equality and Comparison/IsLessThanOrEqualToByKeyPath.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 // MARK: - Is Less than or Equal to by KeyPath
-/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`es.
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
 ///
 ///     isLessThanOrEqual(objectA, to: objectB, by: \.a)
 ///
@@ -23,7 +23,7 @@ public func isLessThanOrEqual<T, P0>(
     !isLess(rhs, than: lhs, by: keyPath0)
 }
 
-/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`es.
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
 ///
 ///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b)
 ///
@@ -40,7 +40,7 @@ public func isLessThanOrEqual<T, P0, P1>(
     !isLess(rhs, than: lhs, by: keyPath0, keyPath1)
 }
 
-/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`es.
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
 ///
 ///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c)
 ///
@@ -59,7 +59,7 @@ public func isLessThanOrEqual<T, P0, P1, P2>(
     !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2)
 }
 
-/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`es.
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
 ///
 ///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c, \.d)
 ///
@@ -80,7 +80,7 @@ public func isLessThanOrEqual<T, P0, P1, P2, P3>(
     !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2, keyPath3)
 }
 
-/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`es.
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
 ///
 ///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c, \.d, \.e)
 ///
@@ -103,7 +103,7 @@ public func isLessThanOrEqual<T, P0, P1, P2, P3, P4>(
     !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2, keyPath3, keyPath4)
 }
 
-/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`es.
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
 ///
 ///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c, \.d, \.e, \.f)
 ///
@@ -126,4 +126,124 @@ public func isLessThanOrEqual<T, P0, P1, P2, P3, P4, P5>(
         P5: Comparable
 {
     !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2, keyPath3, keyPath4, keyPath5)
+}
+
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
+///
+///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c, \.d, \.e, \.f, \.g)
+///
+public func isLessThanOrEqual<T, P0, P1, P2, P3, P4, P5, P6>(
+    _ lhs: T,
+    to rhs: T,
+    by keyPath0: KeyPath<T, P0>,
+    _ keyPath1: KeyPath<T, P1>,
+    _ keyPath2: KeyPath<T, P2>,
+    _ keyPath3: KeyPath<T, P3>,
+    _ keyPath4: KeyPath<T, P4>,
+    _ keyPath5: KeyPath<T, P5>,
+    _ keyPath6: KeyPath<T, P6>
+) -> Bool
+    where
+        P0: Comparable,
+        P1: Comparable,
+        P2: Comparable,
+        P3: Comparable,
+        P4: Comparable,
+        P5: Comparable,
+        P6: Comparable
+{
+    !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2, keyPath3, keyPath4, keyPath5, keyPath6)
+}
+
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
+///
+///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h)
+///
+public func isLessThanOrEqual<T, P0, P1, P2, P3, P4, P5, P6, P7>(
+    _ lhs: T,
+    to rhs: T,
+    by keyPath0: KeyPath<T, P0>,
+    _ keyPath1: KeyPath<T, P1>,
+    _ keyPath2: KeyPath<T, P2>,
+    _ keyPath3: KeyPath<T, P3>,
+    _ keyPath4: KeyPath<T, P4>,
+    _ keyPath5: KeyPath<T, P5>,
+    _ keyPath6: KeyPath<T, P6>,
+    _ keyPath7: KeyPath<T, P7>
+) -> Bool
+    where
+        P0: Comparable,
+        P1: Comparable,
+        P2: Comparable,
+        P3: Comparable,
+        P4: Comparable,
+        P5: Comparable,
+        P6: Comparable,
+        P7: Comparable
+{
+    !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2, keyPath3, keyPath4, keyPath5, keyPath6, keyPath7)
+}
+
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
+///
+///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h, \.i)
+///
+public func isLessThanOrEqual<T, P0, P1, P2, P3, P4, P5, P6, P7, P8>(
+    _ lhs: T,
+    to rhs: T,
+    by keyPath0: KeyPath<T, P0>,
+    _ keyPath1: KeyPath<T, P1>,
+    _ keyPath2: KeyPath<T, P2>,
+    _ keyPath3: KeyPath<T, P3>,
+    _ keyPath4: KeyPath<T, P4>,
+    _ keyPath5: KeyPath<T, P5>,
+    _ keyPath6: KeyPath<T, P6>,
+    _ keyPath7: KeyPath<T, P7>,
+    _ keyPath8: KeyPath<T, P8>
+) -> Bool
+    where
+        P0: Comparable,
+        P1: Comparable,
+        P2: Comparable,
+        P3: Comparable,
+        P4: Comparable,
+        P5: Comparable,
+        P6: Comparable,
+        P7: Comparable,
+        P8: Comparable
+{
+    !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2, keyPath3, keyPath4, keyPath5, keyPath6, keyPath7, keyPath8)
+}
+
+/// Returns a `Bool` value indicating whether the value of the first argument is less than or equal to that of the second argument with given `KeyPath`s.
+///
+///     isLessThanOrEqual(objectA, to: objectB, by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h, \.i, \.j)
+///
+public func isLessThanOrEqual<T, P0, P1, P2, P3, P4, P5, P6, P7, P8, P9>(
+    _ lhs: T,
+    to rhs: T,
+    by keyPath0: KeyPath<T, P0>,
+    _ keyPath1: KeyPath<T, P1>,
+    _ keyPath2: KeyPath<T, P2>,
+    _ keyPath3: KeyPath<T, P3>,
+    _ keyPath4: KeyPath<T, P4>,
+    _ keyPath5: KeyPath<T, P5>,
+    _ keyPath6: KeyPath<T, P6>,
+    _ keyPath7: KeyPath<T, P7>,
+    _ keyPath8: KeyPath<T, P8>,
+    _ keyPath9: KeyPath<T, P9>
+) -> Bool
+    where
+        P0: Comparable,
+        P1: Comparable,
+        P2: Comparable,
+        P3: Comparable,
+        P4: Comparable,
+        P5: Comparable,
+        P6: Comparable,
+        P7: Comparable,
+        P8: Comparable,
+        P9: Comparable
+{
+    !isLess(rhs, than: lhs, by: keyPath0, keyPath1, keyPath2, keyPath3, keyPath4, keyPath5, keyPath6, keyPath7, keyPath8, keyPath9)
 }