+++ Tests/VCoreTests/Tests/Global Functions/KeyPath Equatable and Comparable/IsLessThanByKeyPathTests.swift
@@ -97,4 +97,60 @@ final class IsLessThanByKeyPathTests: XCTestCase {
             by: \.a, \.b, \.c, \.d, \.e, \.f
         ))
     }
+    
+    func test7() {
+        XCTAssertTrue(VCore.isLess(
+            SomeObject(),
+            than: .init(g: 1),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g
+        ))
+        
+        XCTAssertFalse(VCore.isLess(
+            SomeObject(),
+            than: .init(),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g
+        ))
+    }
+    
+    func test8() {
+        XCTAssertTrue(VCore.isLess(
+            SomeObject(),
+            than: .init(h: 1),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h
+        ))
+        
+        XCTAssertFalse(VCore.isLess(
+            SomeObject(),
+            than: .init(),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h
+        ))
+    }
+    
+    func test9() {
+        XCTAssertTrue(VCore.isLess(
+            SomeObject(),
+            than: .init(i: 1),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h, \.i
+        ))
+        
+        XCTAssertFalse(VCore.isLess(
+            SomeObject(),
+            than: .init(),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h, \.i
+        ))
+    }
+    
+    func test10() {
+        XCTAssertTrue(VCore.isLess(
+            SomeObject(),
+            than: .init(j: 1),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h, \.i, \.j
+        ))
+        
+        XCTAssertFalse(VCore.isLess(
+            SomeObject(),
+            than: .init(),
+            by: \.a, \.b, \.c, \.d, \.e, \.f, \.g, \.h, \.i, \.j
+        ))
+    }
 }